MovieAI/movieai/requirements.txt
asgiref==3.8.1
Django==5.2.1
pillow==11.2.1
sqlparse==0.5.3
whitenoise==6.8.2
python-dotenv==1.0.1
